Fix various errors and warnings with Eclipse build.
Due to inclusion of private Eclipse workspace files in GitHub, there
appears to be a synchronization issue with Maven.  The Eclipse files had
what seem to be old, outdated references.

Also fixed an issue in a persistence.xml file where the package name of
some test classes was wrong, probably missed during a refactoring event.

Signed-off-by: Dan Price <g051051+eclipse@gmail.com>